requirements.txt
bcrypt==4.2.0
blinker==1.9.0
click==8.3.1
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
python-dotenv==1.0.1
Werkzeug==3.1.7
